+++ omen/__init__.py
@@ -1,3 +1,8 @@
 '''Init module'''
 
-import main
+from header import *
+
+def init_db(model_name):
+    '''Function for inital db if it not inited'''
+    
+
