--- omen/main.py
@@ -1,5 +1,0 @@
-''' Main module for omen engine '''
-
-# -*- coding: utf-8 -*-
-#Author: Phonphey
-#Slaves: Oksovskii, KDV
